+++ talstrom_server/TalStromApi/Program.cs
@@ -2,13 +2,13 @@
 using System;
 using System.Text.Json.Serialization;
 using Azure.Storage.Blobs;
-using AzureFullstackPractice.Data;
 using AzureFullstackPractice.Services;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.OpenApi.Models;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using TalStromApi.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddDbContext<TalStromDbContext>(options =>
